tcg2: Display UEFI TCG2 capability data.
This commit adds a headers with the structure definitions required for
interacting with the TCG2 UEFI protocol (formerly known as the Microsoft
TrEE protocol). We then add a simple UEFI application that uses this
protocol to invoke the GetCapability function displaying the results on
the console. We also define a new code module to house utility functions
related to the TCG2 protocol.

NOTE: the GetCapability function from the TCG2 UEFI protocol is used to
query data about the TCG2 UEFI protocol driver. This is *NOT* the same
as the TPM2 GetCapability command that is used to query TPM2 devices for
their capabilities.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>