test: Move wrap functions to separate code module.
This required a bunch of refactoring in the way we include the UEFI
types from gnu-efi. The workaround for the conflict between the gnu
setjmp.h and gnu-efi headers is now in its own header file that we
include in the places where we need access to the gnu-efi types. This
include was missing in a few places like the tcg2-protocol.h file.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>