tool: Replace tcti-send-buf with tpm2-get-capability.
The 'tcti-send-buf' tool was largely used to test the 'transmit' and
'receive' functions from the TCTI. Now that we have access to the
libtss2-sys API we can use the Tss2_Sys_GetCapability function. NOTE:
the command buffer sent by the 'tcti-send-buf' tool has the same effect
as invoking the GetCapability function with the parameters used in this
new tool.

The code used to display the data this tool queries for comes from the
tpm2-tools project. Its use of the 'isspace' function caused us a few
problems. It depends on locality stuff in libc that I wasn't able to get
statically compiled in (the __ctype_b_loc symbol). The easiest solution
to this was to provide our own implementation of 'isspace'. This commit
adds 'src/compat.c' for this purpose.

Similarly there are some uses of 'memcpy' in libtss2-mu that I couldn't
get statically compiled into the executable. The 'memcpy' API is
identical to the gnu-efi CopyMem function and so the function added to
the compat module is just a wrapper.

Finally, the output / formatting of the TPM fixed properties is still
pretty rough. Future work will clean this up.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>