ci: Run scan-build on clang.
The most simple way to do this is to abandon the 'compiler' JSON
element. Instead we construct the build matrix manually. By doing this
we can define environment varaibles in the matrix more simply. To invoke
scan-build we use an environment variable that is not defined for the
gcc compiler. This prevents us from having to do tricks with the
'exclude' element.

NOTE: When test cases fail, but scan-build doesn't detect issues the
clang build will succeed. This is a bit less than accurate but the gcc
build will fail regardless and the failed tests caught.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>