build: Add integration test harness using swtpm and qemu + OVMF.
This commit adds a 'log compiler' associated with the '.efi' file
suffix. When the test harness encounters a file with this suffix in the
TESTS variable it executes this new script 'lib/efi-test-setup.sh'. This
script does the following:
- Create a temp directory for the test context
- Moves the test executable to this directory (required for qemu / OVMF
to launch the test)
- Generate a 'startup.nsh' script that causes OVMF to automatically
launch the test executable, output a status message on success / fail
and to shutdown qemu once the test is done.
- Launch swtpm to provide TPM2 backend for the test
- Launch qemu redirecting all output to the console and ensuring qemu is
shutdown after a 30 second timeout (necessary since broken tests may
cause qemu to hang indefinitely).

Additionally this patch moves the 'hello-world' and 'tss2-mu-uint32' to
the test directory where all integration tests will go.

NOTE: We do not enable the integration tests in travis-ci here because
the version of qemu / OVMF shipped with travis does not work with swtpm.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>