Add dependencies for integration tests to build-deps.sh & enable in CI.
These dependencies are required to run the integration tests on the
Ubuntu instance used by travis-ci (xenial). A more current distro (like
Debian testing / buster) will be able to install them from packages.

This commit removes the use of the gnu-efi packaged in the travis-ci
build image building it the 3.0.9 source release instead. We require a
version > 3.0.8.

Clang won't build the master branch of swtpm currently. This commit
includes a workaround that forces the use of gcc for this component
only.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>