build: Add docker build file based on Ubuntu xenial.
Use this file to build a docker image complete with all build
dependencies. Xenial was chosen for the base image because it is the
closest thing to a travis-ci builder we have.

To build the image and use it to build and test the tpm2-tcti-uefi
sources execute following from the root of the tpm2-tcti-uefi source
tree:

$ docker image build \
    --file ./Dockerfile-xenial \
    --tag tpm2-tcti-uefi:builder $(pwd)

Then run a container using the image built above mapping in the
tpm2-tcti-uefi sources (assumed in pwd) to /src:

$ docker run -d -it ${PWD}:/src tpm2-tcti-uefi:builder

Then build the tpm2-tcti-uefi sources and run the tests:

$ docker exec -it <container-id> /bin/bash -c "cd /src && make distcheck"

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>