ci: Build OVMF from source.
The images previously obtained from kraxel.org are automated weekly
builds. They rotate weekly and so using them is more complex than just
getting the URL right. Building OVMF from source is more reliable.

This commit adds the necessary dependencies to .travis.yml. Building
OVMF is a bit more complicated:
- Firstly the EDK2 build requires bash and so the build-deps.sh script
now uses bash in the shebang.
- The EDK2 build requires a configuration file that we store under
$(srcdir)/lib and copy into the cloned EDK2 sources.
- The compiler used is listed in this file and so we force the use of
GCC for simplicity.
- The way that travis-ci configures python3 is a problem for the EDK2
build and so we disable it in favor of python2.7.
- Finally we find the OVMF.fd file and copy it into place.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>