ci: Enable code coverage when building with GCC & send to coveralls.
We only collect code coverage when building with GCC. This requires a
new environment variable to hold additional options passed to the
configure script (--enable-code-coverage). Also we can't use the
distcheck target to collect code coverage metrics so we add yet another
environment variable so that we can use 'check' with gcc. To collect
these metrics we need lcov installed as well. Finally we upload metrics
to coveralls only if CC is gcc.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>