configure.ac: tweak usage of AC_ARG_ENABLE
For AC_ARG_ENABLE(x, y, ACTION-IF-FOUND, ACTION-IF-NOT-FOUND),
when ACTION-IF-FOUND and ACTION-IF-NOT-FOUND are not provided, that is
AC_ARG_ENABLE(x, y) is in configure.ac, and
  “./configure” is called, the variable enable_x is not set,
                that is ${enable_x+set} != set;
  "./configure --enable-x" is called, the variable enable_x is set to "yes";
  "./configure --disable-x" is called, the variable enable_x is set to "no".

For features disabled by default, the only thing that matters is, if the user
calls “./configure --enable-x”, that is, if enable_x is "yes".  enable_x being
not set, or being "no" is the same state.

On “./configure --help” printing “--enable-x” implies that feature X is
disabled by default and the help instructs the user how to enable the feature.
Printing --disable-X would imply the opposite.

Signed-Off-By: Дилян Палаузов <git-dpa@aegee.org>