example: Add support for the "crypto agile" TPM event log format.
This format is significantly more complex than the older 1.2 version.
The old version had only a single fixed length SHA1 hash with a variable
length 'event' blob. Events in this new format include multiple,
variable length hashes as well as the event blob. This makes extracting
the various fields a bit more complicated.

The output format is pretty close to the format used for the 1.2
eventlog.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>