Force gnu-efi and edk2 headers to coexist.
We use autoconf to ensure that we have the required gnuefi headers when
building on Linux. When building under edk2 we don't have a
configuration phase and so we can't detect attributes of the build
environment before building. edk2 provides only static build metadata
files where we can set compiler / linker flags.

This patch introduces a new macro that the edk2 build metadata passes
to the compiler to indicate edk2 headers should be used (EDK2_BUILD).
We wrap the include directives for these headers in preprocessor
if/else/endif. If this macro is not defined then we assume we're
building under gnuefi.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>