ci: build gnuefi 3.0.10
This is required to use EFIAPI macro in tss2-tcti-uefi.h. This requires
including efi.h from gnu-efi which pulled in the gnu-efi implementation
of setjmp. This conflicts with the system version of setjmp used by
cmocka. The gnu-efi headers were fixed by simply removing the inclusion
of their setjmp implementation since it's not needed by efi.h.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>